requirements.txt
asgiref==3.7.2
Django==5.0.2
django-colorfield==0.11.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.2.0
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2024.1
